Make possible to customize znode path
