+++ pkg/serversets/interfaces.go
@@ -2,6 +2,7 @@ package serversets
 
 // ZKRecord defines the APIs an endpoint record in zookeeper should provide.
 type ZKRecord interface {
+	// Marshal encodes data to be stored in zookeeper.
 	Marshal() ([]byte, error)
 	// Endpoint returns the endpoint as host:port string
 	Endpoint() string
@@ -9,8 +10,14 @@ type ZKRecord interface {
 	IsAlive() bool
 }
 
-// ZKRecordProvider defines the APIs for a endpoints' zookeeper record provider.
-type ZKRecordProvider interface {
+// ZKFmt defines the APIs for endpoint's zookeeper path and data formats.
+type ZKFmt interface {
+	// Path returns the zokeeper full path to service members.
+	Path() string
+	// Prefix returns the znode name prefix for service members.
+	Prefix() string
+	// CreateRecord creates a new endpoint record from host and port.
 	Create(host string, port int) ZKRecord
+	// Unmarshal decodes encoded endpoint data a ZKrecord.
 	Unmarshal([]byte) (ZKRecord, error)
 }
